Improve accessibility (#4)
Improve accessibility to reach lighthouse 100% score and axe dev tools 0 issues.

The main changes are to headings (h1 > h2), layout (header > main etc.), contrast (medium grey on dark gery to lighter gray on dark grey), adjusting padding to allow for focus rings and add aria-labels.

There is still some issues, for example, the tab order is incorrect when using keyboard navigation to tab through the site on mobile view. You will tab from the header, to the footer, then back up to the body. The fix is to render the alternative nav/footer in two locations using a component, the alternative nav is hidden on mobile and the footer hidden on tablets (md).

I hope this helps you and others.